Move `Translation` to `translation` module.
Although it started as an offshoot of `SliceExt`, it feels really
odd for it not to be in `translation`.
